+++ zad2Extended/zad2Extended/ClassB.cs
@@ -10,11 +10,11 @@ namespace zad2Extended
     public class ClassB : ISerializable
     {
         public String name { get; set; }
-        public int a { get; set; }
+        public float a { get; set; }
         public bool iftrue { get; set; }
         public ClassC classC { get; set; }
         public DateTime data { get; set; }
-        public ClassB(String name, int a, bool iftrue,DateTime data, ClassC classC)
+        public ClassB(String name, float a, bool iftrue,DateTime data, ClassC classC)
         {
             this.a = a;
             this.name = name;
@@ -31,8 +31,9 @@ namespace zad2Extended
 
             this.name = info.GetString("nazwa");
             this.iftrue = Boolean.Parse(info.GetString("iftrue"));
-            this.a = Int32.Parse(info.GetString("wartosc"));
+            this.a = Single.Parse(info.GetString("wartosc"));
             this.classC = null;
+            
 
         }
         public void GetObjectData(SerializationInfo info, StreamingContext context)
